Add gest_sale_id helper
